Duration always displayed as hh:mm:ss

+++ tulip-runtime/src/main/kotlin/io/github/wfouche/tulip/core/Tulip.kt
@@ -404,7 +404,7 @@ private data class ActionSummary(
 
     var histogram: Histogram = Histogram(histogramNumberOfSignificantValueDigits),
 
-    var durationSeconds: Double = 0.0,
+    var durationSeconds: Int = 0,
 
     var aps: Double = 0.0,
     var art: Double = 0.0,
@@ -470,10 +470,10 @@ private class ActionStats {
         r.numActions = numActions
         r.numSuccess = numSuccess
 
-        r.durationSeconds = durationMillis.toDouble() / 1000.0
+        r.durationSeconds = durationMillis.toInt() / 1000
 
         // actions per second (aps)
-        r.aps = numActions / r.durationSeconds
+        r.aps = numActions / r.durationSeconds.toDouble()
 
         // average response time (art) in milliseconds
         r.art = histogram.mean / 1000.0
